tidy(ui): remove unused files
