+++ invokeai/frontend/web/src/app/store/middleware/listenerMiddleware/listeners/enqueueRequestedLinear.ts
@@ -1,8 +1,8 @@
 import { enqueueRequested } from 'app/store/actions';
 import type { AppStartListening } from 'app/store/middleware/listenerMiddleware';
 import { isImageViewerOpenChanged } from 'features/gallery/store/gallerySlice';
-import { buildGenerationTabGraph2 } from 'features/nodes/util/graph/buildGenerationTabGraph2';
-import { buildGenerationTabSDXLGraph2 } from 'features/nodes/util/graph/buildGenerationTabSDXLGraph2';
+import { buildGenerationTabGraph } from 'features/nodes/util/graph/buildGenerationTabGraph';
+import { buildGenerationTabSDXLGraph } from 'features/nodes/util/graph/buildGenerationTabSDXLGraph';
 import { prepareLinearUIBatch } from 'features/nodes/util/graph/buildLinearBatchConfig';
 import { queueApi } from 'services/api/endpoints/queue';
 
@@ -18,10 +18,10 @@ export const addEnqueueRequestedLinear = (startAppListening: AppStartListening) 
 
       let graph;
 
-      if (model && model.base === 'sdxl') {
-        graph = await buildGenerationTabSDXLGraph2(state);
+      if (model?.base === 'sdxl') {
+        graph = await buildGenerationTabSDXLGraph(state);
       } else {
-        graph = await buildGenerationTabGraph2(state);
+        graph = await buildGenerationTabGraph(state);
       }
 
       const batchConfig = prepareLinearUIBatch(state, graph, prepend);
